requirements.txt
requests
SQLAlchemy
PyMySQL
six
python-dotenv
praw
beautifulsoup4